created dialogs.py added browsing remote filesystem
